requirements.txt
# Django 기본 패키지
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-environ>=0.10.0  # 환경 변수 관리

# 데이터베이스
psycopg2-binary>=2.9.9  # PostgreSQL 드라이버

# API 문서 (Swagger UI)
drf-yasg>=1.21.7

# 인증 및 보안
djangorestframework-simplejwt>=5.3.0  # JWT 인증
bcrypt>=4.1.2  # 비밀번호 암호화

# 비동기 작업 처리
celery>=5.3.4
redis>=5.0.1  # Redis 캐싱 및 Celery 백엔드

# CORS 및 요청 관리
django-cors-headers>=4.3.1
djangorestframework-camel-case>=1.3.0  # JSON 응답을 camelCase로 변환

# 파일 업로드 및 스토리지
django-storages>=1.14.2  # AWS S3 등 외부 저장소 사용 가능

# 배포 및 성능 최적화
gunicorn>=21.2.0  # WSGI 서버 (배포 시 사용)
whitenoise>=6.6.0  # 정적 파일 서빙 최적화
uvicorn>=0.27.1  # ASGI 서버 (FastAPI, WebSocket 지원)

# 개발 도구
python-dotenv>=1.0.1  # 환경 변수 로드
ipython>=8.18.1  # Django shell+에서 유용

# 테스트 및 디버깅
pytest>=8.0.2
pytest-django>=4.8.0
factory-boy>=3.3.0  # 테스트 데이터 생성

# Docker 지원
dj-database-url>=2.1.0  # DATABASE_URL 환경 변수 지원